requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
Pillow==9.3.0
sqlparse==0.4.3

djangorestframework~=3.14.0